deps/v8/test/inspector/debugger/regress-crbug-1319828-expected.txt

Regression test for crbug.com/1319828

Running test: testDebuggerGetPossibleBreakpoints

function foo() {
  |_|console.|C|log('Hello World!');
|R|}


Running test: testDebuggerSetBreakpointByUrl
function foo() {
  #console.log('Hello World!');
}

function foo() {
  console.#log('Hello World!');
}
